+++ docs/components/content/examples/RangeExample.vue
@@ -1,7 +1,7 @@
-<script setup>
+<script setup lang="ts">
 const value = ref(50)
 </script>
 
 <template>
-  <URange v-model="value" name="range" />
+  <URange v-model="value" />
 </template>