+++ src/runtime/components/elements/Meter.vue
@@ -31,6 +31,7 @@
 import { computed, defineComponent, toRef } from 'vue'
 import type { SlotsType, PropType } from 'vue'
 import { twJoin } from 'tailwind-merge'
+import UIcon from './Icon.vue'
 import { useUI } from '../../composables/useUI'
 import { mergeConfig } from '../../utils'
 import type { Strategy, MeterColor, MeterSize } from '../../types'
@@ -41,6 +42,9 @@ import { meter } from '#ui/ui.config'
 const config = mergeConfig<typeof meter>(appConfig.ui.strategy, appConfig.ui.meter, meter)
 
 export default defineComponent({
+  components: {
+    UIcon
+  },
   inheritAttrs: false,
   slots: Object as SlotsType<{
     indicator?: { percent: number, value: number },