+++ src/runtime/types/horizontal-navigation.d.ts
@@ -1,0 +1,13 @@
+import type { Link } from './link'
+import type { Avatar } from './avatar'
+import type { Badge } from './badge'
+
+export interface HorizontalNavigationLink extends Link {
+  label: string
+  labelClass?: string
+  icon?: string
+  iconClass?: string
+  avatar?: Avatar
+  click?: Function
+  badge?: string | number | Badge
+}