+++ src/runtime/types/tooltip.d.ts
@@ -1,0 +1,7 @@
+export interface Tooltip {
+  text?: string
+  prevent?: boolean
+  shortcuts?: string[]
+  openDelay?: number
+  closeDelay?: number
+}