+++ src/runtime/ui.config/navigation/verticalNavigation.ts
@@ -1,8 +1,8 @@
 export default {
   wrapper: 'relative',
-  base: 'group relative flex items-center gap-2 focus:outline-none focus-visible:outline-none dark:focus-visible:outline-none focus-visible:before:ring-inset focus-visible:before:ring-1 focus-visible:before:ring-primary-500 dark:focus-visible:before:ring-primary-400 before:absolute before:inset-px before:rounded-md disabled:cursor-not-allowed disabled:opacity-75',
+  base: 'group relative flex items-center gap-1.5 focus:outline-none focus-visible:outline-none dark:focus-visible:outline-none focus-visible:before:ring-inset focus-visible:before:ring-1 focus-visible:before:ring-primary-500 dark:focus-visible:before:ring-primary-400 before:absolute before:inset-px before:rounded-md disabled:cursor-not-allowed disabled:opacity-75',
   ring: 'focus-visible:ring-inset focus-visible:ring-2 focus-visible:ring-primary-500 dark:focus-visible:ring-primary-400',
-  padding: 'px-3 py-1.5',
+  padding: 'px-2.5 py-1.5',
   width: 'w-full',
   rounded: 'rounded-md',
   font: 'font-medium',
@@ -11,17 +11,23 @@ export default {
   inactive: 'text-gray-500 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white hover:before:bg-gray-50 dark:hover:before:bg-gray-800/50',
   label: 'truncate relative',
   icon: {
-    base: 'flex-shrink-0 w-4 h-4',
+    base: 'flex-shrink-0 w-5 h-5 relative',
     active: 'text-gray-700 dark:text-gray-200',
     inactive: 'text-gray-400 dark:text-gray-500 group-hover:text-gray-700 dark:group-hover:text-gray-200'
   },
   avatar: {
     base: 'flex-shrink-0',
-    size: '3xs' as const
+    size: '2xs' as const
   },
   badge: {
-    base: 'relative ms-auto inline-block py-0.5 px-2 text-xs rounded-md -me-1 -my-0.5',
-    active: 'bg-white dark:bg-gray-900',
-    inactive: 'bg-gray-100 dark:bg-gray-800 text-gray-900 dark:text-white group-hover:bg-white dark:group-hover:bg-gray-900'
+    base: 'flex-shrink-0 ml-auto relative rounded',
+    color: 'gray' as const,
+    variant: 'solid' as const,
+    size: 'xs' as const
+  },
+  divider: {
+    wrapper: {
+      base: 'p-2'
+    }
   }
 }